Change int/uint to isize/usize

+++ src/sdl2/cpuinfo.rs
@@ -1,13 +1,13 @@
 pub use sys::cpuinfo as ll;
 
-pub const CACHELINESIZE: int = 128;
+pub const CACHELINESIZE: isize = 128;
 
-pub fn get_cpu_count() -> int {
-    unsafe { ll::SDL_GetCPUCount() as int }
+pub fn get_cpu_count() -> isize {
+    unsafe { ll::SDL_GetCPUCount() as isize }
 }
 
-pub fn get_cpu_cache_line_size() -> int {
-    unsafe { ll::SDL_GetCPUCacheLineSize() as int}
+pub fn get_cpu_cache_line_size() -> isize {
+    unsafe { ll::SDL_GetCPUCacheLineSize() as isize}
 }
 
 pub fn has_rdtsc() -> bool {
@@ -50,6 +50,6 @@ pub fn has_avx() -> bool {
     unsafe { ll::SDL_HasAVX() == 1 }
 }
 
-pub fn get_system_ram() -> int {
-    unsafe { ll::SDL_GetSystemRAM() as int }
+pub fn get_system_ram() -> isize {
+    unsafe { ll::SDL_GetSystemRAM() as isize }
 }
