+++ src/sdl2/keycode.rs
@@ -260,7 +260,7 @@ impl ToPrimitive for KeyCode {
     }
 
     #[inline]
-    fn to_int(&self) -> Option<int> {
-        Some(*self as int)
+    fn to_int(&self) -> Option<isize> {
+        Some(*self as isize)
     }
 }
